services/pii-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.26.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
pytest>=8.0
pytest-asyncio>=0.23.0
